refactored teams -> teams header and teams filter
